+++ Gems/ROS2/Code/ros2_files.cmake
@@ -64,6 +64,7 @@ set(FILES
         Source/Spawner/ROS2SpawnerComponent.h
         Source/Spawner/ROS2SpawnPointComponent.cpp
         Source/Spawner/ROS2SpawnPointComponent.h
+        Source/Utilities/Controllers/PidConfiguration.cpp
         Source/Utilities/ROS2Conversions.cpp
         Source/Utilities/ROS2Names.cpp
         Source/VehicleDynamics/AxleConfiguration.cpp
@@ -74,7 +75,6 @@ set(FILES
         Source/VehicleDynamics/DriveModels/AckermannDriveModel.h
         Source/VehicleDynamics/DriveModels/SkidSteeringDriveModel.cpp
         Source/VehicleDynamics/DriveModels/SkidSteeringDriveModel.h
-        Source/VehicleDynamics/DriveModels/PidConfiguration.cpp
         Source/VehicleDynamics/ManualControlEventHandler.h
         Source/VehicleDynamics/Utilities.cpp
         Source/VehicleDynamics/Utilities.h
